+++ Masonry/MASConstraint.h
@@ -120,6 +120,18 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, readonly) MASConstraint *equalToSuperview;
 
 /**
+ *	Sets the constraint relation to NSLayoutRelationGreaterThanOrEqual with the superview
+ *  This is equivalent to greaterThanOrEqualTo(self.superview)
+ */
+@property (nonatomic, readonly) MASConstraint *greaterThanOrEqualToSuperview;
+
+/**
+ *	Sets the constraint relation to NSLayoutRelationLessThanOrEqual with the superview
+ *  This is equivalent to lessThanOrEqualTo(self.superview)
+ */
+@property (nonatomic, readonly) MASConstraint *lessThanOrEqualToSuperview;
+
+/**
  *	Optional semantic property which has no effect but improves the readability of constraint
  */
 @property (nonatomic, readonly) MASConstraint *with;