+++ Masonry/MASConstraint.m
@@ -66,6 +66,18 @@
     return self.equalToWithRelation(superview, NSLayoutRelationEqual);
 }
 
+- (MASConstraint *)greaterThanOrEqualToSuperview {
+    id superview = [self mas_superview];
+    NSAssert(superview != nil, @"greaterThanOrEqualToSuperview requires the constrained item to have a superview.");
+    return self.equalToWithRelation(superview, NSLayoutRelationGreaterThanOrEqual);
+}
+
+- (MASConstraint *)lessThanOrEqualToSuperview {
+    id superview = [self mas_superview];
+    NSAssert(superview != nil, @"lessThanOrEqualToSuperview requires the constrained item to have a superview.");
+    return self.equalToWithRelation(superview, NSLayoutRelationLessThanOrEqual);
+}
+
 #pragma mark - MASLayoutPriority proxies
 
 - (MASConstraint * (^)(void))priorityLow {